test next

+++ app/src/main/java/de/hdmstuttgart/wetter/datenmodellForecastApiCall/WeatherDataForecast.kt
@@ -11,7 +11,7 @@ data class WeatherDataForecast (
 
   @SerializedName("cod") var cod: String? = null,
   @SerializedName("message") var message: Int? = null,
-  @SerializedName("cnt") var cnt: Int? = null,
+  @SerializedName("cnt") var cnt: Double? = null,
   @SerializedName("list") var list: ArrayList<WeatherList> = arrayListOf(),
   @SerializedName("city") var city: City? = City()
